+++ crates/zk-commit-core/src/lib.rs
@@ -3,3 +3,4 @@ pub mod types;
 pub mod circuit_config;
 pub mod claim_circuit;
 pub mod example;
+pub mod circuit_utils;
